Commit tasks on Mod+Enter
